refactor: 로그인, 마이페이지 불필요한 코드 개선작업
